fix upscale_factor bug

+++ config.py
@@ -10,7 +10,7 @@ parser.add_argument('--sample_output_folder', type=str, default='samples',  help
 parser.add_argument('--use_gpu', type=str2bool, default=True, help='Use GPU')
 parser.add_argument('--gpu', type=int, default=-1, help='GPU id')
 parser.add_argument('--lr', type=float, default=1e-4, help='Learning Rate. Default=1e-4')
-parser.add_argument('--upscale_factor', type=float, default=2, help='upscale_factor. Default=2')
+parser.add_argument('--upscale_factor', type=int, default=2, help='upscale_factor. Default=2')
 parser.add_argument('--num_workers', type=int, default=4, help='Number of threads for data loader, for window set to 0')
 parser.add_argument('--print_gap', type=int, default=50, help='number of batches to print average loss ')
 parser.add_argument('--batch_size', type=int, default=16, help='Training batch size')
